API Reference

Types

BlockSparseMatrices.BlockSparseMatrix — Type

struct BlockSparseMatrix{T,M,D} <: AbstractBlockMatrix{T}

Sparse block matrix.

Fields

blocks::Vector{M}: Dense blocks.

size::Tuple{Int,Int}: Size of the matrix.

forwardbuffer::Vector{T}: ???.

adjointbuffer::Vector{T}: ???.

buffer::Vector{T}: ???.

rowindexdict::D: Global rowindices of dense blocks.

colindexdict::D: Global columnindices of dense blocks.

threadsafecolors::Vector{Vector{Int}}: Coloring of dense blocks for multithreaded matrix-vector product.

ntasks::Int: Threads used for the matrix Vector product.

BlockSparseMatrices.SymmetricBlockMatrix — Type

SymmetricBlockMatrix{T,DM,M,D} <: AbstractBlockMatrix{T}

Symmetric sparse block matrix.

Fields

diagonals::Vector{DM}: Dense blocks on the diagonal of the matrix.

offdiagonals::Vector{M}: Dense blocks not on the diagonal.

size::Tuple{Int,Int}: Size of the matrix.

forwardbuffer::Vector{T}: ???.

adjointbuffer::Vector{T}: ???.

buffer::Vector{T}: ???.

diagonalrowindexdict::D: Global rowindices of diagonal dense blocks.

diagonalcolindexdict::D: Global columnindices of offdiagonal dense blocks.

offdiagonalrowindexdict::D: Global rowindices of diagonal dense blocks.

offdiagonalcolindexdict::D: Global columnindices of offdiagonal dense blocks.

threadsafecolors::Vector{Vector{Int}}: Coloring of dense blocks for multithreaded matrix-vector product.

ntasks::Int: Threads used for the matrix Vector product.

BlockSparseMatrices.DenseMatrixBlock — Type

DenseMatrixBlock{T,M,RC} <: AbstractMatrixBlock{T}

Dense block in a sparse blockmatrix.

Fields

matrix::M: Matrix.

rowindices::RC: Global row indices.

colindices::RC: Global column indices.

Functions

BlockSparseMatrices.findcolor! — Function

findcolor!(
    blockid::Int,
    threadsafecolors::AbstractArray,
    blocks::Vector{M};
    threadsafecheck=isthreadsafe(),
    color=1,
) where {M}

Assigns recursively threadsafe color to the dense block with index blockid.

Arguments

blockid::Int: Index of block.

threadsafecolors::AbstractArray: Contains assigned blocks.

blocks::Vector{M}: Vector with all blocks.

threadsafecheck=isthreadsafe(): Threadsafe check, differs between symmetric and non-symmetric matrix.

color=1: Currently tested color, increased if block does not fit in color.

BlockSparseMatrices.islessinordering — Function

islessinordering(blocka::AbstractMatrixBlock, blockb::AbstractMatrixBlock)

Sorting rule for AbstractMatrixBlock objects